feat(settings): add TimePickerDialog composable for settings
* Introduced a reusable TimePickerDialog for better time selection in settings
* Simplified dropdown menu configurations in BankAccountsScreen and NavGraph
* Adjusted box height calculation in MonthTrendGraph for accurate rendering

+++ app/src/main/java/com/personalfinance/tracker/ui/screens/MonthTrendGraph.kt
@@ -195,7 +195,7 @@ fun MonthTrendGraph(
                 val textStyle = TextStyle(fontSize = 11.sp, color = Color.White)
                 val measured = lines.map { measurer.measure(AnnotatedString(it), textStyle) }
                 val boxW = (measured.maxOfOrNull { it.size.width } ?: 0) + 20.dp.toPx()
-                val boxH = measured.sumOf { (it.size.height + 4.dp.toPx()).toDouble() } + 12.dp.toPx()
+                val boxH = measured.sumOf { (it.size.height + 4.dp.toPx()).toDouble() }.toFloat() + 12.dp.toPx()
 
                 val groupX = sidePadPx + idx * groupW
                 var boxX = groupX + groupW / 2 - boxW / 2
